Speedup, ~20%
